feat: Restructure packages to streamline the addition of new algorithms
