+++ docs/source/examples.rst
@@ -6,4 +6,3 @@ Here can find different Jupyter notebook examples about how to use `bdi-kit`:
 - `Changing  the parameters of the matching methods <https://github.com/VIDA-NYU/bdi-kit/blob/devel/examples/changing_parameters.ipynb>`__
 - `Getting the top-k value matches <https://github.com/VIDA-NYU/bdi-kit/blob/devel/examples/top_k_value_matches.ipynb>`__
 - `Analyzing one attribute/column at a time <https://github.com/VIDA-NYU/bdi-kit/blob/devel/examples/analyzing_one_attribute.ipynb>`__
-
